backend/requirements.txt
asgiref==3.8.1
boto3==1.37.13
botocore==1.37.13
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
dj-database-url==2.3.0
Django==5.1.7
django-anymail==12.0
django-ckeditor-5==0.2.17
django-cors-headers==4.7.0
django-jazzmin==3.0.1
django-storages==1.14.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
environs==14.1.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
marshmallow==3.26.1
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.11.4
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
stripe==11.6.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0